requirements-common.txt
pyyaml==5.1
wheel==0.33.1
ipython==7.5
lxml==4.2.5
html5lib==1.0.1
urllib3==1.24.1
Cython==0.29.4
gunicorn==19.9.0
psycopg2-binary==2.8.2
falcon==2.0.0
jsonschema<3.0
click<7.0
werkzeug==0.14.1
flatdict==3.1.0
marshmallow==3.0.0rc1
apispec==1.3.3
webargs==5.3.1
pyjwt==1.7.1
fleep==1.0.1
querystring-parser==1.2.3
python-magic==0.4.15
argon2-cffi==19.1.0
pillow==5.4.1
rules==2.0
beautifulsoup4==4.6.3
bs4==0.0.1
fire==0.1.3
celery==4.2.1
ezodf==0.3.2
rdflib==4.2.2
tabulator==1.20.0
tabulator[ods]==1.20.0
elasticsearch-dsl==6.1.0
goodtables==2.1.4
goodtables[ods]==2.1.4
tqdm==4.28.1
deepdiff==3.3.0
pygments==2.3.0
passlib==1.7.1
markdown2==2.3.5
marshmallow-jsonapi==0.21.0
json-api-doc==0.7.0
redis==2.10.6
openapi-core==0.7.1
gevent==1.2.2
phonenumbers==8.10.5
django==2.2
django-constance[database]==2.3.1
django-elasticsearch-dsl==0.5.1
django-redis==4.10
django-suit==0.2.26
django-model-utils==3.1.2
django-localflavor==2.1
django-environ==0.4.5
django-ckeditor==5.6.1
django-nested-admin==3.1.2
django-autocomplete-light==3.3.2
django-admin-rangefilter==0.3.12
dal_admin_filters==0.2.2
django-celery-results==1.0.1
django-celery-beat==1.1.1
django-cache-memoize==0.1.6
django-tqdm==0.0.3
django-modeltrans==0.3.4
dj-database-url==0.5.0
boto3==1.9.112
django-vault-helpers==0.8.0
changelog-cli==0.6.2
libarchive-c==2.8
elastic-apm==4.2.1
python-logstash==0.4.6
python-mimeparse==1.6.0
pyshp==2.1.0
